Commit Text Organizer Package Tests: *m test_argument_parser.py *m test_commit_line.py *m test_commit_text_group.py *m test_commit_text_group_challenges.py *m test_commit_line_prefixes.py *m test_commit_organizer.py *m test_file_validation.py
