+++ tests/test_commit_line_group_challenges.py
@@ -1,6 +1,6 @@
 import unittest
 
-from commit_text_organizer.commit_text_group import *
+from commit_text_organizer.commit_line_group import *
 
 input_text_challenge_2 = tuple("""
 * Third Challenge-a,  c
@@ -21,6 +21,6 @@ class TestCommitTextGroupChallenges(unittest.TestCase):
 		print([str(x.get_subject()) for x in lines])
 		print([str(x.get_content()) for x in lines])
 
-		cto = CommitTextGroup(None, input_text_challenge_2)
+		cto = CommitLineGroup(None, input_text_challenge_2)
 		cto.autoprocess()
 		print(str(cto))
